+++ urllib3/backends/__init__.py
@@ -4,6 +4,6 @@ from .sync_backend import SyncBackend
 __all__ = ['SyncBackend']
 
 if six.PY3:
-    from .._async.backends.trio_backend import TrioBackend
-    from .._async.backends.twisted_backend import TwistedBackend
+    from .trio_backend import TrioBackend
+    from .twisted_backend import TwistedBackend
     __all__ += ['TrioBackend', 'TwistedBackend']
