+++ urllib3/backends/trio_backend.py
@@ -1,6 +1,6 @@
 import trio
 
-from ...backends._common import is_readable, LoopAbort
+from ._common import is_readable, LoopAbort
 
 BUFSIZE = 65536
